+++ demo/extras/src/main/java/it/gmariotti/cardslib/demo/extras/fragment/twowayview/StaggeredGrid2wayFragment.java
@@ -43,7 +43,8 @@ import it.gmariotti.cardslib.demo.extras.staggered.data.ServerDatabase;
 import it.gmariotti.cardslib.library.extra.twoway.view.CardTwowayView;
 import it.gmariotti.cardslib.library.internal.Card;
 import it.gmariotti.cardslib.library.internal.CardThumbnail;
-import it.gmariotti.cardslib.library.internal.recyclerview.CardArrayRecyclerViewAdapter;
+import it.gmariotti.cardslib.library.recyclerview.internal.CardArrayRecyclerViewAdapter;
+
 
 /**
  * This example uses a staggered card with different different photos and text.