+++ library-recyclerview/src/main/java/it/gmariotti/cardslib/library/recyclerview/internal/ViewAdapterImpl.java
@@ -16,7 +16,7 @@
  *  *****************************************************************************
  */
 
-package it.gmariotti.cardslib.library.internal.recyclerview;
+package it.gmariotti.cardslib.library.recyclerview.internal;
 
 import android.support.annotation.NonNull;