+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 2.0.0 *(2014-XX-XX)*
 ----------------------------
 * LIB: Added the new **CardViewNative**![(Learn how to migrate your cards)](/doc/MIGRATE)
 * LIB: New aar modules. [(doc)[(BUILD.md#including-in-your-project-with-gradle)
+* LIB-RECYCLERVIEW: Ner module with support for the new RecyclerView
 * LIB: Added a CardRecyclerView and a CardArrayRecyclerViewAdapter [(doc)](/doc/CARDRECYCLERVIEW.md)
 * LIB-CARDS: New module with built-in cards.
 * LIB-CARDS: Material Cards [(doc)](/doc/MATERIALCARD.md)